+++ dutch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.29
+
+- Sync Dutch 1.3.27 source to upstream commit `d4e96f60108d`.
+
 ## 1.3.28
 
 - Sync Dutch 1.3.26 source to upstream commit `bd88bbb4cf0a`.